Update tests related to task listing
